agents/shared/python/tests/test_jwt.py
"""JWT verifier smoke tests.

These cover the verifier's local failure modes — bad kid, expired
token, wrong issuer, wrong audience — without needing a live Cognito
JWKS endpoint. A full-path test with a real RS256 key pair lives in
the integration suite (out of scope for this public portfolio build).
"""

from __future__ import annotations

import base64
import json
import time
from unittest.mock import MagicMock

import pytest
from jose.exceptions import JWTError

import sys
import pathlib

sys.path.insert(0, str(pathlib.Path(__file__).resolve().parents[1] / "src"))
from jwt_verifier import JwtVerifier  # noqa: E402


def _segment(payload: dict) -> str:
    raw = json.dumps(payload, separators=(",", ":")).encode()
    return base64.urlsafe_b64encode(raw).rstrip(b"=").decode()


def _fake_token(*, kid: str, payload: dict) -> str:
    header = _segment({"alg": "RS256", "kid": kid, "typ": "JWT"})
    body = _segment(payload)
    # Dummy signature — we're testing the pre-signature code paths.
    return f"{header}.{body}.AAAA"


def _build_verifier(http_mock: MagicMock | None = None) -> JwtVerifier:
    return JwtVerifier(
        user_pool_id="us-east-1_POOL",
        region="us-east-1",
        client_id="client-xyz",
        http_client=http_mock,
    )


def test_missing_kid_raises():
    v = _build_verifier()
    token = "e30.e30.AAAA"  # header: {}, payload: {}
    with pytest.raises(JWTError, match="Missing kid"):
        v.verify(token)


def test_unknown_kid_triggers_refresh_and_raises():
    http = MagicMock()
    http.get.return_value.raise_for_status.return_value = None
    http.get.return_value.json.return_value = {"keys": []}  # empty after refresh
    v = _build_verifier(http)
    token = _fake_token(kid="nope", payload={})
    with pytest.raises(JWTError, match="Unknown kid"):
        v.verify(token)
    assert http.get.called


def test_expired_token_raises():
    # Populate the verifier's cache with a stub "key" that verify() can't
    # use — the test stops at the signature-verify step, which is fine
    # because the before-signature paths still need to catch expiration.
    v = _build_verifier()
    past = int(time.time()) - 3600
    token = _fake_token(kid="kid-1", payload={"exp": past, "iss": "x"})
    # First put a key in the cache so the flow reaches signature verify.
    v._jwks_by_kid["kid-1"] = {
        "kid": "kid-1",
        "kty": "RSA",
        "use": "sig",
        "alg": "RS256",
        "n": "x",
        "e": "AQAB",
    }
    # The construct() call will fail with our stub modulus. That's fine
    # for this test: we're just asserting the pre-signature guards run.
    with pytest.raises(Exception):  # noqa: PT011 — any failure past header parsing is ok
        v.verify(token)
