agents/shared/python/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-sip-shared"
version = "0.1.0"
description = "Shared Python runtime utilities for ai-sales-intelligence-platform agent services."
requires-python = ">=3.12"
dependencies = [
  "boto3>=1.34",
  "psycopg[binary,pool]>=3.1",
  "structlog>=24.1",
  "python-jose[cryptography]>=3.3",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "httpx>=0.27",
  "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"
